requirements.txt
shiny>=1.0
shinywidgets>=0.3
plotly>=5.18
pandas>=2.0
openpyxl>=3.1
scikit-learn>=1.2
matplotlib>=3.7
seaborn>=0.12
statsmodels>=0.14
scipy>=1.10
fredapi>=0.4
python-dotenv>=0.19
supabase>=2.0